LibGfx: Delete DeprecatedPainter
