+++ Libraries/LibGfx/Forward.h
@@ -22,13 +22,10 @@ class ScaledFont;
 template<typename T>
 class Line;
 
-class AntiAliasingPainter;
-class DeprecatedPainter;
 class Painter;
 class PaintingSurface;
 class Palette;
 class PaletteImpl;
-class DeprecatedPath;
 class Path;
 class ShareableBitmap;
 struct SystemTheme;
